Add root directory to be used by tests.
On Mac, the `temp_dir` function, often returns the same path -
`/var/folders/...` for multiple test runs.
This means that consecutive tests may interfere.
To address this, a `root_dir` fixture was added in the main test module.
This fixture may then be used in other test modules.
Crate `chrono` and `rstest` was added to the project.